Move AppInterface definition
