+++ src/definition/metadata/index.ts
@@ -3,6 +3,8 @@ import { IAppAuthorInfo } from './IAppAuthorInfo';
 import { IAppInfo } from './IAppInfo';
 import { RocketChatAssociationModel, RocketChatAssociationRecord } from './RocketChatAssociations';
 
+export * from './AppInterface';
+
 export {
     AppMethod,
     IAppAuthorInfo,
